2D plot now shows the received object positions
And the plotly version was bumped up.
It does seem like there is an error of ~2cm between the computed position and the position that the robot targets. (The computed position is computed in the same way as the robot would compute it).

Perhaps this is due to the old issue of the Calculated TCP position being a little wrong