Improved the scrolling smoothness, removing the stuttering
Previously a line update while a scroll is in progress, makes the scroll stop. This has been changed by throttling the calls that trigger a scroll. The color updates are still being triggered as often as before, but the scrolling only happens every 600ms (which is definitely okay, if a scroll is requested at say 500, it is put into a queue, which has a capacity of 1, i find it quite nice)